feat: rename ProfileDto to ProfileDetailDto and add ProfileDto
