+++ httpRequests/internal-profile-search.http
@@ -1,0 +1,5 @@
+### GET request to example server
+GET https://examples.http-client.intellij.net/get$END$
+    ?generated-in=IntelliJ IDEA
+
+###
